Add conditional to removal migration for customer battle tag and skype

+++ db/migrate/20151229233001_drop_battle_tag_from_customers.rb
@@ -1,6 +1,11 @@
 class DropBattleTagFromCustomers < ActiveRecord::Migration
   def change
-    remove_column :customers, :battle_tag
-    remove_column :customers, :skype
+    if column_exists?(:customers, :battle_tag)
+      remove_column :customers, :battle_tag
+    end
+    
+    if column_exists?(:customers, :skype)
+      remove_column :customers, :skype
+    end
   end
 end
